+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added `assert_error_covers`.
 - Add more logs (gh-326).
 - Add `justrun` helper as a tarantool runner and output catcher (gh-365).
+- Changed error message for too long Unix domain socket paths (gh-341).
 
 ## 1.0.1